requirements.txt
# Core AI/ML
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.35.0
accelerate>=0.24.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0

# Web Interface
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.0

# Style Transfer & Enhancement
kornia>=0.7.0
lpips>=0.1.4
real-esrgan>=0.3.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
omegaconf>=2.3.0
python-dotenv>=1.0.0

# Optional: For advanced features
open-clip-torch>=2.0.0
invisible-watermark>=0.2.0